molresponse: extract shared DALTON AO->MRA projection into tools/dalton_mra.hpp
The Gaussian-LC functor + occ-vir block projector lived as private copies in
tpa_from_dalton.cpp and seed_from_dalton.cpp; the dalton.dir FD seed path
(next commit) is a third consumer, so hoist them into a header and point both
tools at it. read_molden additionally records the [Atoms] atomic numbers —
the geometry fingerprint keys element identity on them, not on the free-form
atom labels (DALTON writes O_a / H_b style tags).

No behavior change intended for the tools; seed_from_dalton's block projector
now scales the AO weights before projection instead of scaling the projected
function (linear either way; truncation now sees the scaled function).
